FEATURE: Add global and worker timeout option.
This PR addresses the feedback what was originally given on PR #162 by
adding a default worker specific timeout option in addition to a global
timeout.

The original commit

https://github.com/discourse/image_optim/commit/8bf3c0e32f50cd2e3e23c0b5b936947994ff75e8

on PR #162 with just the global timeout has been in production since
2018-07-08 on Discourse instances using a forked version of this gem.

We would like to get this change merged in, so that we can get off of
the discourse specific fork and use the latest version of the
image_optim gem.

Co-authored-by: Blake Erickson <o.blakeerickson@gmail.com>