Do not raise if optimized before timeout
Based on feedback this commit includes the following changes:

When optimizing multiple images with `optimize_images` a single image
timeout doesn't break execution. The global timeout will still raise an
exception though if it is reached.

When optimizing a single image do not raise an exception if a single
worker has succeeded but has reached the timeout period while running
through the remaining workers.

When optimizing a single image if a single worker times out it will
continue on to the next worker until the global timeout is reached.